Move out custom methods in data-classes to syntax package
These are mostly unused, so we might be able to get rid of them too eventually?
