+++ modules/definitions/src/main/scala/lmcoursier/credentials/DirectCredentials.scala
@@ -15,16 +15,8 @@ import dataclass._
   httpsOnly: Boolean = true
 ) extends Credentials {
 
-  def withRealm(realm: String): DirectCredentials =
-    withRealm(Option(realm))
-
   override def toString(): String =
     withPassword("****")
       .productIterator
       .mkString("DirectCredentials(", ", ", ")")
-}
-
-object DirectCredentials {
-  def apply(host: String, username: String, password: String, realm: String): DirectCredentials = DirectCredentials(host, username, password, Option(realm))
-  def apply(host: String, username: String, password: String, realm: String, optional: Boolean): DirectCredentials = DirectCredentials(host, username, password, Option(realm), optional)
 }
