+++ modules/definitions/src/main/scala/lmcoursier/definitions/ModuleMatchers.scala
@@ -11,12 +11,3 @@ import dataclass.data
   include: Set[Module],
   includeByDefault: Boolean = true
 )
-
-object ModuleMatchers {
-  def all: ModuleMatchers =
-    ModuleMatchers(Set.empty, Set.empty)
-  def only(organization: String, moduleName: String): ModuleMatchers =
-    ModuleMatchers(Set.empty, Set(Module(Organization(organization), ModuleName(moduleName), Map())), includeByDefault = false)
-  def only(mod: Module): ModuleMatchers =
-    ModuleMatchers(Set.empty, Set(mod), includeByDefault = false)
-}
