+++ modules/definitions/src/main/scala/lmcoursier/definitions/Publication.scala
@@ -7,10 +7,4 @@ import dataclass.data
   `type`: Type,
   ext: Extension,
   classifier: Classifier
-) {
-  def attributes: Attributes =
-    Attributes(`type`, classifier)
-  def withAttributes(attributes: Attributes): Publication =
-    withType(attributes.`type`)
-      .withClassifier(attributes.classifier)
-}
+)
